requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.3.2
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.9
Django==5.1.5
dotenv==0.9.9
filelock==3.18.0
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.53.1
gunicorn==23.0.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
joLib==0.0.1
kiwisolver==1.4.5
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.2
nltk==3.9.1
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pandas-datareader==0.10.0
pandas-io==0.0.1
pillow==11.1.0
platformdirs==4.3.7
pyparsing==3.1.4
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
regex==2024.11.6
requests==2.32.3
setuptools==74.0.0
six==1.16.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.30.0
Werkzeug==3.1.3
whitenoise==6.9.0
